+++ examples/common/bakerAdd.ts
@@ -3,16 +3,15 @@ import {
     AccountTransaction,
     AccountTransactionHeader,
     AccountTransactionType,
-    buildBasicAccountSigner,
     signTransaction,
     TransactionExpiry,
     createConcordiumClient,
     CcdAmount,
-    unwrap,
-    HexString,
     generateBakerKeys,
     ConfigureBakerPayload,
     OpenStatus,
+    parseWallet,
+    buildAccountSigner,
 } from '@concordium/node-sdk';
 import { credentials } from '@grpc/grpc-js';
 import { readFileSync } from 'node:fs';
@@ -68,11 +67,9 @@ const client = createConcordiumClient(
 
 (async () => {
     // Read wallet-file
-    const walletFile = JSON.parse(readFileSync(cli.flags.walletFile, 'utf8'));
-    const sender = new AccountAddress(unwrap(walletFile.value.address));
-    const senderPrivateKey: HexString = unwrap(
-        walletFile.value.accountKeys.keys[0].keys[0].signKey
-    );
+    const walletFile = readFileSync(cli.flags.walletFile, 'utf8');
+    const wallet = parseWallet(walletFile);
+    const sender = new AccountAddress(wallet.value.address);
 
     const header: AccountTransactionHeader = {
         expiry: new TransactionExpiry(new Date(Date.now() + 3600000)),
@@ -100,7 +97,7 @@ const client = createConcordiumClient(
     };
 
     // Sign transaction
-    const signer = buildBasicAccountSigner(senderPrivateKey);
+    const signer = buildAccountSigner(wallet);
     const signature = await signTransaction(
         configureBakerAccountTransaction,
         signer
@@ -112,6 +109,7 @@ const client = createConcordiumClient(
     );
 
     console.log('Transaction submitted, waiting for finalization...');
+
     await client.waitForTransactionFinalization(transactionHash);
 
     console.log('Transaction finalized, getting outcome...\n');