+++ packages/common/CHANGELOG.md
@@ -21,6 +21,9 @@
     - `findInstanceCreation`
     - `findFirstFinalizedBlockNoLaterThan`
 
+### Breaking changes
+- `waitForTransactionFinalization` now returns a `BlockItemSummary`
+
 ## 6.5.0 2023-5-03
 
 ### Added